refactor(profile, nft): simplify favorites flow and extract shared UI components
- Refactor favorites routing:
  - Change ProfileRoute.favorites to use [String] instead of Profile
  - Update ProfileView and TabBarView to use favorite NFT ids

- Update FavoriteNFTsViewModel:
  - Replace Profile dependency with favoriteIds
  - Add async updateFavoriteIds closure for persistence

- Add ProfileViewModel.updateFavoriteIds to save updated likes
- Add Profile.updatingLikes helper for immutable profile updates

- Extract reusable NFT UI components:
  - FavoriteNFTImageView
  - RatingStarsView
- Remove duplicated implementations and redundant Kingfisher imports

- Update related views/components:
  - FavoriteNFTsView
  - FavoriteNFTCell
  - MyNFTsRowView
  - ProfileView
  - TabBarView

- Simplify preview setup and mock usage
