+++ iOS-FakeNFT-Extended/UI/Profile/Views/FavoriteNFTCell.swift
@@ -5,7 +5,6 @@
 //  Created by Анастасия Федотова on 19.05.2026.
 //
 
-import Kingfisher
 import SwiftUI
 
 struct FavoriteNFTCell: View {
@@ -55,7 +54,7 @@ struct FavoriteNFTCell: View {
                 .foregroundStyle(Color.ypBlack)
                 .lineLimit(1)
 
-            FavoriteRatingStarsView(rating: nft.rating ?? 0)
+            RatingStarsView(rating: nft.rating ?? 0)
                 .padding(.top, 4)
 
             Text(priceText)
@@ -83,112 +82,6 @@ struct FavoriteNFTCell: View {
         }
 
         return String(format: "%.2f ETH", locale: Locale(identifier: "ru_RU"), price)
-    }
-}
-
-// MARK: - FavoriteNFTImageView
-
-private struct FavoriteNFTImageView: View {
-
-    // MARK: - State
-
-    @State private var didFail = false
-
-    // MARK: - Properties
-
-    let url: URL?
-
-    // MARK: - Body
-
-    var body: some View {
-        ZStack {
-            if let url, !didFail {
-                KFImage(url)
-                    .placeholder {
-                        ZStack {
-                            Color.ypBackgroundUniversal
-
-                            LoadingSpinner(size: .medium, tint: .ypWhiteUniversal)
-                        }
-                    }
-                    .retry(maxCount: 2, interval: .seconds(1))
-                    .fade(duration: 0.2)
-                    .onFailure { _ in
-                        didFail = true
-                    }
-                    .resizable()
-                    .scaledToFill()
-            } else {
-                FavoriteNFTImagePlaceholder()
-            }
-        }
-        .frame(width: 80, height: 80)
-        .clipped()
-    }
-}
-
-// MARK: - FavoriteNFTImagePlaceholder
-
-private struct FavoriteNFTImagePlaceholder: View {
-
-    // MARK: - Constants
-
-    private enum Constants {
-        static let cellSize: CGFloat = 20
-        static let gridSize = 4
-    }
-
-    // MARK: - Body
-
-    var body: some View {
-        VStack(spacing: 0) {
-            ForEach(0..<Constants.gridSize, id: \.self) { row in
-                HStack(spacing: 0) {
-                    ForEach(0..<Constants.gridSize, id: \.self) { column in
-                        Rectangle()
-                            .fill((row + column).isMultiple(of: 2) ? Color.ypWhite : Color.ypBlack)
-                            .frame(width: Constants.cellSize, height: Constants.cellSize)
-                    }
-                }
-            }
-        }
-        .frame(width: 80, height: 80)
-    }
-}
-
-// MARK: - FavoriteRatingStarsView
-
-private struct FavoriteRatingStarsView: View {
-
-    // MARK: - Constants
-
-    private enum Constants {
-        static let maxRating = 5
-        static let starSize: CGFloat = 12
-        static let totalWidth: CGFloat = 60
-    }
-
-    // MARK: - Properties
-
-    let rating: Int
-
-    private var normalizedRating: Int {
-        min(max(rating, 0), Constants.maxRating)
-    }
-
-    // MARK: - Body
-
-    var body: some View {
-        HStack(spacing: 0) {
-            ForEach(1...Constants.maxRating, id: \.self) { index in
-                Image(index <= normalizedRating ? .starActive : .starInactive)
-                    .renderingMode(.original)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: Constants.starSize, height: Constants.starSize)
-            }
-        }
-        .frame(width: Constants.totalWidth, height: Constants.starSize, alignment: .leading)
     }
 }
